+++ forms-flow-api/src/formsflow_api/schemas/filter.py
@@ -42,6 +42,7 @@ class FilterSchema(AuditDateTimeSchema):
     isMyTasksEnabled = fields.Bool(load_only=True)
     isTasksForCurrentUserGroupsEnabled = fields.Bool(load_only=True)
     order = fields.Int(data_key="order", allow_none=True)
+    sort_order = fields.Int(data_key="sortOrder", allow_none=True, dump_only=True)
     filter_type = fields.Method(
         "get_filter_type",
         deserialize="load_filter_type",